renderer: split gltf environment and renderer responsibilities
